+++ src/design/plone/contenttypes/restapi/deserializers/dxfields.py
@@ -9,6 +9,9 @@ from plone.restapi.deserializer.dxfields import DefaultFieldDeserializer
 from plone.formwidget.geolocation.geolocation import Geolocation
 from design.plone.contenttypes import _
 from zope.i18n import translate
+from design.plone.contenttypes.fields import IBlocksField
+from zope.component import subscribers
+from plone.restapi.interfaces import IBlockFieldDeserializationTransformer
 
 
 @implementer(IFieldDeserializer)
@@ -26,4 +29,30 @@ class GeolocationFieldDeserializer(DefaultFieldDeserializer):
                     context=self.request,
                 )
             )
-        return Geolocation(latitude=value["latitude"], longitude=value["longitude"])
+        return Geolocation(
+            latitude=value["latitude"], longitude=value["longitude"]
+        )
+
+
+@implementer(IFieldDeserializer)
+@adapter(IBlocksField, IDexterityContent, IBrowserRequest)
+class BlocksFieldDeserializer(DefaultFieldDeserializer):
+    def __call__(self, value):
+        value = super(BlocksFieldDeserializer, self).__call__(value)
+        blocks = value.get("blocks", {})
+        if blocks:
+            for id, block_value in blocks.items():
+                block_type = block_value.get("@type", "")
+                handlers = []
+                for h in subscribers(
+                    (self.context, self.request),
+                    IBlockFieldDeserializationTransformer,
+                ):
+                    if h.block_type == block_type or h.block_type is None:
+                        handlers.append(h)
+                for handler in sorted(handlers, key=lambda h: h.order):
+                    if not getattr(handler, "disabled", False):
+                        block_value = handler(block_value)
+
+                blocks[id] = block_value
+        return value